Select a physical device
Selection is based on a simple scoring system where the bare minimum is
that the device has the geometry shader feature, and that it has a
graphics queue. Discrete GPUs are then favored over integrated GPUs, and
in turn all other Vulkan capable devices. Finally, the largest 2D image
that the device is capable of is taken into account.