Implement a staging buffer to put vertex data in
The idea behind using a staging buffer is that the GPU memory that we
can map from the host is not the most optimal memory for use on the GPU.
Therefore we create a staging buffer that we map from the host and that
we can put our vertex data in. We then copy the data into the device
local memory which should be more performant for later use.